silence compiler warnings

+++ ext/linenoise/linenoise.cpp
@@ -479,7 +479,7 @@ class Utf32String {
   Utf32String& operator=(const Utf32String& that) {
     if (this != &that) {
       delete[] _data;
-      _data = new (std::nothrow) char32_t[that._length + 1]();
+      _data = new (std::nothrow) char32_t[static_cast<uint32_t>(that._length) + 1]();
       _length = that._length;
       memcpy(_data, that._data, sizeof(char32_t) * _length);
     }
